refactor: rename tool.ts to agent tool/utils.ts in ai service
