+++ apps/backend/src/services/ai/agent/tool/utils.ts
@@ -1,4 +1,4 @@
-import { AgentToolFactory } from "./types.js";
+import { AgentToolFactory } from "../types.js";
 import { Tool } from "@openai/agents";
 
 /**
